Group the cut-off date logic together
